<doc> Write product requirements document
Used Gemini integration with google doc to write the PRD document. The
document was then converted to markdown format. It is still barebones as
many details still need to be fleshed out.

The main functional and non-functional requirements are written down.
Specifics will be filled in later as the product develops.