<feat> Move postgres config to .env file
Commit @724efb used hardcoded connection string to connect to postgres
database. That is not good practice and thus it was moved to a .env
file.

The connection String is built using information read from the .env file
which contains the following keys:
    1. POSTGRES_HOST
    2. POSTGRES_PORT
    3. POSTGRES_USER
    4. POSTGRES_PASSWORD
    5. POSTGRES_DB
    6. POSTGRES_SSL_MODE
    7. POSTGRES_DRIVER

This ensures that later on the database can be changed without having to
change the code.