<feat> Fetch array of deck details
Added functionality to query and fetch an array containing details of
all decks in the database. However, this includes only the following
columns:
- id
- name
- description
- total_cards

This is deliberate because it will be used when displaying the list of
decks to the user in the home page. The other columns will only be
displayed when the user clicks on a specific deck to view its details.
This decision was made to optimize the data transfer.