<bugfix> Adjusted unit test from commit acbf07
Unit test for GET /deck endpoint, TestGetAllDecksHandlerNonEmpty,
runs into an error. The cause is due to randomness assocaited with
map data structure. As it doesn't return items in sorted order. This
results in the test sometimes passing and sometimes failing.

Instead of checking returned body string, lengths are checked instead.