<feat> Show deck details as a card
Card shows name, description, creation date, modification date, and total cards.
Last studied date is not shown as their is a bug in the backend. The last stuy date
should be null if the deck has never been studied. But it has been set to
a very old date to avoid null errors. This should be updated.