<feat> Get total cards /deck/{id}/card/total
This commit adds a new endpoint to the API to get the total number of cards in a deck.
A design decision was made to return 0 for decks that don't exist.
This was done to simplify the backend logic. Otherwise, I'd have to handle
the case where a deck doesn't exist in the database and throw error accordingly.